Update for Fastify v5 (#347)
* update

* remove coverage check

* update ci
