+++ .taprc
@@ -1,5 +1,2 @@
-ts: false
-jsx: false
-flow: false
-coverage: true
-check-coverage: true
+files:
+  - test/**/*.js
