+++ package.json
@@ -28,34 +28,34 @@
   },
   "homepage": "https://github.com/fastify/fastify-http-proxy#readme",
   "devDependencies": {
-    "@fastify/pre-commit": "^2.0.2",
-    "@fastify/websocket": "^9.0.0",
-    "@types/node": "^20.1.0",
-    "@types/ws": "^8.2.2",
-    "@typescript-eslint/eslint-plugin": "^7.1.0",
-    "@typescript-eslint/parser": "^7.1.0",
-    "express": "^4.17.2",
+    "@fastify/pre-commit": "^2.1.0",
+    "@fastify/websocket": "^10.0.1",
+    "@types/node": "^20.12.7",
+    "@types/ws": "^8.5.10",
+    "@typescript-eslint/eslint-plugin": "^7.6.0",
+    "@typescript-eslint/parser": "^7.6.0",
+    "express": "^4.19.2",
     "express-http-proxy": "^2.0.0",
     "fast-proxy": "^2.1.0",
-    "fastify": "^4.0.0-rc.2",
-    "got": "^11.8.3",
+    "fastify": "^4.26.2",
+    "got": "^11.8.6",
     "http-errors": "^2.0.0",
     "http-proxy": "^1.18.1",
-    "simple-get": "^4.0.0",
+    "simple-get": "^4.0.1",
     "snazzy": "^9.0.0",
-    "socket.io": "^4.4.1",
-    "socket.io-client": "^4.4.1",
-    "standard": "^17.0.0",
-    "tap": "^16.0.0",
+    "socket.io": "^4.7.5",
+    "socket.io-client": "^4.7.5",
+    "standard": "^17.1.0",
+    "tap": "^18.7.2",
     "tsd": "^0.31.0",
-    "typescript": "^5.0.2",
+    "typescript": "^5.4.5",
     "why-is-node-running": "^2.2.2"
   },
   "dependencies": {
-    "@fastify/reply-from": "^9.0.0",
+    "@fastify/reply-from": "^9.7.0",
     "fast-querystring": "^1.1.2",
-    "fastify-plugin": "^4.5.0",
-    "ws": "^8.4.2"
+    "fastify-plugin": "^4.5.1",
+    "ws": "^8.16.0"
   },
   "tsd": {
     "directory": "test/types"
